+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Switched to using `<FormattedUTCDate>` from Stripes. ERM-635
 * Switched to using `<Spinner>` from Stripes. ERM-635
 * Keyboard navigation and focus improvements. ERM-593 622 628
+* Added @folio/react-intl-safe-html to dependencies. ERM-539
 
 ## 3.6.0 2019-12-04
 * Port ui-licenses to RFF. ERM-487.